requirements.txt
click>= 8.0.3
# cudatoolkit>= 10.2
faiss-gpu
matplotlib>= 3.5.0
pillow>= 8.4.0
pretrainedmodels>= 0.7.4
# torch>= 1.10.0
scikit-image>= 0.18.3
scikit-learn>= 1.0.1
scipy>= 1.7.1
# torchvision>= 0.11.1
tqdm>= 4.62.3
timm 
opencv-python
imgaug 
tensorboard 
numpy 
Pillow
wandb
